% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{addLegend}
\alias{addLegend}
\title{Adds legend to the plot (used internally by plotDeletion)}
\usage{
addLegend(cand, trBlack, mainText = "")
}
\description{
Adds legend to the plot (used internally by plotDeletion)
}
