% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{annotateAOH}
\alias{annotateAOH}
\title{Finds overlap between calls and AOH regions}
\usage{
annotateAOH(candidatesMergedAnnotated, extAOH, aohSize, minAOHsig, mc.cores)
}
\arguments{
\item{candidatesMergedAnnotated}{object returned by annotateCandidates}

\item{extAOH}{object returned by prepareAOHData}

\item{aohSize}{min AOH size}

\item{minAOHsig}{AOH signal threshold}

\item{mc.cores}{number of cores (see mclapply2)}
}
\description{
Finds overlap between calls and AOH regions
}
