% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{annotateCandidates}
\alias{annotateCandidates}
\title{Adds information about internal sample name and project (BHCMG specific)}
\usage{
annotateCandidates(candidatesMerged, is_cmg = FALSE)
}
\arguments{
\item{candidatesMerged}{object returned by mergeCandidates}
}
\description{
Adds information about internal sample name and project (BHCMG specific)
}
