% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{annotateFreq}
\alias{annotateFreq}
\title{Adds the information about overlapping calls (OverlapCnt column) to the list of candidates}
\usage{
annotateFreq(filteredCalls)
}
\arguments{
\item{filteredCalls}{object returned by annotateAOH}
}
\description{
Adds the information about overlapping calls (OverlapCnt column) to the list of candidates
}
