% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{calcRPKMsFromBAMs}
\alias{calcRPKMsFromBAMs}
\title{Calculate RPKMs from BAM files
NOTE: Index files (.bai) should be available for each BAM file}
\usage{
calcRPKMsFromBAMs(bedFile, bamFiles, sampleNames, outputDir, mc.cores)
}
\arguments{
\item{bedFile}{tab separeated bed file containing capture design probes with the following 4 columns:
- chromosme
- start
- stop
- gene name}

\item{bamFiles}{vector of paths to bam files}

\item{sampleNames}{vector of sample identifiers}

\item{outputDir}{output directory}

\item{mc.cores}{number of cores}
}
\description{
Calculate RPKMs from BAM files
NOTE: Index files (.bai) should be available for each BAM file
}
