% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{calculateZscores}
\alias{calculateZscores}
\title{Calculates average z-RPKM for each call}
\usage{
calculateZscores(filteredCalls, rpkmDtOrdered)
}
\description{
Calculates average z-RPKM for each call
}
