% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{getCandidateExonCalls}
\alias{getCandidateExonCalls}
\title{Retrieves potential HMZ deletions}
\usage{
getCandidateExonCalls(rpkmDtOrdered, bedOrdered, selectedExonsFinal, mc.cores,
  lowRPKMthreshold)
}
\arguments{
\item{rpkmDtOrdered}{rpkmDtOrdered object returned by reorderBedAndRpkmDt}

\item{bedOrdered}{bedOrdered object returned by reorderBedAndRpkmDt}

\item{selectedExonsFinal}{object returned by processRPKM}

\item{mc.cores}{number of cores (see mclapply2)}

\item{lowRPKMthreshold}{RPKM threshold}
}
\description{
Retrieves potential HMZ deletions
}
