% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{mergeCandidates}
\alias{mergeCandidates}
\title{Merges calls from consecutive exons}
\usage{
mergeCandidates(candidateExonCalls, bedOrdered, maxGap = 10)
}
\arguments{
\item{candidateExonCalls}{object returned by getCandidateExonCalls}

\item{bedOrdered}{bedOrdered object returned by reorderBedAndRpkmDt}

\item{maxGap}{maximal distance between deleted exons to be merged}
}
\description{
Merges calls from consecutive exons
}
