% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{plotDeletion}
\alias{plotDeletion}
\title{Plots read depth for selected candidate of HMZ deletion}
\usage{
plotDeletion(calls, i, bedOrdered, rpkmDtOrdered, lowRPKMthreshold, outputDir,
  mainText = "")
}
\arguments{
\item{calls}{calls object returned by runHMZDelFinder}

\item{i}{index of the deletion to plot}

\item{bedOrdered}{bedOrdered object returned by runHMZDelFinder}

\item{rpkmDtOrdered}{rpkmDtOrdered object returned by runHMZDelFinder}

\item{outputDir}{output directory}
}
\description{
Plots read depth for selected candidate of HMZ deletion
}
