% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{prepareAOHData}
\alias{prepareAOHData}
\title{Reads VCF files and generates AOH data using CBS}
\usage{
prepareAOHData(fileNames, fids, mc.cores, vR_id, tR_id, filter)
}
\arguments{
\item{fileNames}{a vector of paths to VCF files}

\item{fids}{a vector of  sample names of the same length as fileName}

\item{mc.cores}{number of cores (see mclapply)}
}
\description{
Reads VCF files and generates AOH data using CBS
}
