% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{prepareRPKMData}
\alias{prepareRPKMData}
\title{Reads RPKM files and creates rpkmDt data.table with
samples in rows and exons in columns}
\usage{
prepareRPKMData(fileNames, fids, mc.cores)
}
\arguments{
\item{fileNames}{a vector of paths to RPKM files}

\item{fids}{a vector of sample names of the same length as fileNames}

\item{mc.cores}{number of cores (see mclapply)}
}
\description{
Reads RPKM files and creates rpkmDt data.table with
samples in rows and exons in columns
}
