% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{processRPKM}
\alias{processRPKM}
\title{Selects exons with potential HMZ deletions}
\usage{
processRPKM(rpkmDtOrdered, bedOrdered, mc.cores, lowRPKMthreshold,
  exonsToExclude, maxFrequency = 0.005)
}
\arguments{
\item{rpkmDtOrdered}{rpkmDtOrdered object returned by reorderBedAndRpkmDt}

\item{bedOrdered}{bedOrdered object returned by reorderBedAndRpkmDt}

\item{mc.cores}{number of cores to be used}

\item{lowRPKMthreshold}{RPKM threshold}

\item{maxFrequency}{max frequency of HMZ deletion}
}
\description{
Selects exons with potential HMZ deletions
}
