% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{reorderBedAndRpkmDt}
\alias{reorderBedAndRpkmDt}
\title{Reorder BED and RPKM matrices}
\usage{
reorderBedAndRpkmDt(bedFile, rpkmDt)
}
\arguments{
\item{bedFile}{tab separeated bed file containing capture design probes with the following 4 columns:
- chromosme
- start
- stop
- gene name}

\item{rpkmDt}{object returned by prepareRPKMData}
}
\description{
Reorder BED and RPKM matrices
}
