% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMZDelFinder.R
\name{runHMZDelFinder}
\alias{runHMZDelFinder}
\title{Main function: return HMZ deletion calls}
\usage{
runHMZDelFinder(snpPaths, snpFids, rpkmPaths, rpkmFids, mc.cores, aohRDataOut,
  bedFile, lowRPKMthreshold, minAOHsize, minAOHsig, is_cmg, vR_id, tR_id,
  filter)
}
\arguments{
\item{snpPaths}{list of paths to VCFs with SNP data; if NULL algorithm skips AOH analysis}

\item{snpFids}{list of sample identifiers (in the same order as in snpFileNames); if NULL algorithm skips AOH analysis}

\item{rpkmFids}{list of sample identifiers (in the same order as in rpkmPaths)}

\item{mc.cores}{number of cores}

\item{aohRDataOut}{temporary file that store AOH data}

\item{bedFile}{target design file}

\item{lowRPKMthreshold}{RPKM threshold used in the algorithm (default=0.65)}

\item{minAOHsize}{minimal size of AOH region (default=1000)}

\item{minAOHsig}{threshold for calling AOH (default=0.45)}

\item{is_cmg}{CMG specific annotations (default=FALSE)}

\item{vR_id}{ID for 'the number of variants reads' in VCF FORMAT column (default='VR')}

\item{tR_id}{ID for 'the number of total reads' in VCF FORMAT column (default='DP')}

\item{filter}{only variants with this value in the VCF FILTER column will be used in AOH analysis}

\item{rpkmPath}{list of paths to RPKM data}
}
\description{
Main function: return HMZ deletion calls
}
